refactor : repackage auth
