+++ src/main/java/app/auth/controller/LoginController.java
@@ -1,11 +1,11 @@
-package app.auth;
+package app.auth.controller;
 
+import app.auth.service.LoginService;
+import app.auth.dto.TokenRequest;
+import app.auth.dto.TokenResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
