+++ src/main/java/app/auth/service/UserDetail.java
@@ -1,4 +1,4 @@
-package app.auth;
+package app.auth.service;
 
 public class UserDetail {
     private Long id;
