+++ src/main/java/app/auth/support/AuthenticationException.java
@@ -1,4 +1,4 @@
-package app.auth;
+package app.auth.support;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
