+++ src/main/java/app/auth/support/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package app.auth;
+package app.auth.support;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
