+++ src/main/java/app/auth/util/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package app.auth;
+package app.auth.util;
 
 import io.jsonwebtoken.*;
 import org.springframework.beans.factory.annotation.Value;
