+++ src/test/java/app/nextstep/theme/ThemeE2ETest.java
@@ -2,8 +2,8 @@ package app.nextstep.theme;
 
 import io.restassured.RestAssured;
 import app.nextstep.AbstractE2ETest;
-import app.auth.TokenRequest;
-import app.auth.TokenResponse;
+import app.auth.dto.TokenRequest;
+import app.auth.dto.TokenResponse;
 import app.nextstep.member.MemberRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
